fix dublication by creating a hook
