+++ src/containers/my-courses/course-toolbar/CourseToolbar.tsx
@@ -1,10 +1,4 @@
-import {
-  useCallback,
-  useMemo,
-  SyntheticEvent,
-  ChangeEvent,
-  FocusEvent
-} from 'react'
+import { useCallback, SyntheticEvent, ChangeEvent, FocusEvent } from 'react'
 import { useTranslation } from 'react-i18next'
 import FormControl from '@mui/material/FormControl'
 import Box from '@mui/material/Box'
@@ -19,6 +13,7 @@ import FormHelperText from '@mui/material/FormHelperText'
 
 import { subjectService } from '~/services/subject-service'
 import { categoryService } from '~/services/category-service'
+import useUserCategoriesAndSubjects from '~/hooks/use-user-categories-and-subjects'
 import AppTextField from '~/components/app-text-field/AppTextField'
 import AppToolbar from '~/components/app-toolbar/AppToolbar'
 import DividedDropdownAutocomplete from '~/components/divider-dropdown-autocomplete/DividerDropdownAutocomplete'
@@ -29,12 +24,10 @@ import {
   CourseForm,
   ProficiencyLevelEnum,
   TextFieldVariantEnum,
-  CourseAutocompleteOptionsEnum,
   ComponentEnum,
   UserResponse,
   CourseExtendedAutocompleteOptions
 } from '~/types'
-import { customOptions } from '~/utils/course-custom-options'
 import { styles } from '~/containers/my-courses/course-toolbar/CourseToolbar.style'
 
 interface CourseToolbarProps {
@@ -62,9 +55,10 @@ const CourseToolbar = ({
   handleNonInputValueChange
 }: CourseToolbarProps) => {
   const { t } = useTranslation()
+  const { transformCategories, transformSubjects } =
+    useUserCategoriesAndSubjects({ user })
 
   const { category, subject, proficiencyLevel } = data
-  const { Subjects } = CourseAutocompleteOptionsEnum
   const levelLists = Object.values(ProficiencyLevelEnum)
 
   const getSubjectsNames = useCallback(
@@ -104,28 +98,6 @@ const CourseToolbar = ({
     <FormHelperText error>
       {t('common.errorMessages.proficiencyLevel')}
     </FormHelperText>
-  )
-
-  const userCategories = useMemo(
-    () => user?.mainSubjects.tutor.map((item) => item.category.name) ?? [],
-    [user?.mainSubjects.tutor]
-  )
-
-  const userSubjects = useMemo(
-    () => user?.mainSubjects.tutor.map((item) => item.name) ?? [],
-    [user?.mainSubjects.tutor]
-  )
-
-  const transformCategories = useCallback(
-    (response: CategoryNameInterface[]): CourseExtendedAutocompleteOptions[] =>
-      customOptions(response, userCategories),
-    [userCategories]
-  )
-
-  const transformSubjects = useCallback(
-    (response: SubjectNameInterface[]): CourseExtendedAutocompleteOptions[] =>
-      customOptions(response, userSubjects, Subjects),
-    [Subjects, userSubjects]
   )
 
   const AppAutoCompleteList = (
